Move "read.py" into geost.io; add layout for other types
