+++ geost/__init__.py
@@ -1,5 +1,5 @@
 from geost import data
-from geost.read import (
+from geost.io.read import (
     get_bro_objects_from_bbox,
     get_bro_objects_from_geometry,
     read_borehole_table,
