+++ geost/io/read.py
@@ -1,8 +1,7 @@
 from pathlib import Path
-from typing import Any, Callable
+from typing import Any
 
 import geopandas as gpd
-import numpy as np
 import pandas as pd
 from pyproj import CRS
